requirements.txt
discord.py
requests 
spotipy